+++ src/raster/mod.rs
@@ -6,7 +6,7 @@ mod types;
 mod warp;
 
 pub use rasterband::{Buffer, ByteBuffer, ColorInterpretation, RasterBand, ResampleAlg};
-pub use rasterize::{rasterize, BurnSource, MergeAlgorithm, RasterizeOptions};
+pub use rasterize::{rasterize, BurnSource, MergeAlgorithm, OptimizeMode, RasterizeOptions};
 pub use types::{GDALDataType, GdalType};
 pub use warp::reproject;