+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Make `gdal::rasters::OptimizeMode` public
+  - <https://github.com/georust/gdal/pull/224>
 
 ## 0.8 - 0.10