+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Make `gdal::rasters::OptimizeMode` public
+  - <https://github.com/georust/gdal/pull/224>
+
 ## 0.8 - 0.10
 
 - Update types to fix build on ppc64le.